src/backend/requirements.txt
python-dotenv
asyncio
uvicorn
fastapi[standard]
requests
azure-monitor-opentelemetry
azure-cosmos
azure-identity
aiohttp